@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.7",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.3",
-    "@angular/common": "^7.2.7",
-    "@angular/compiler": "^7.2.7",
-    "@angular/core": "^7.2.7",
-    "@angular/forms": "^7.2.7",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.2.7",
-    "@angular/platform-browser": "^7.2.7",
-    "@angular/platform-browser-dynamic": "^7.2.7",
-    "@angular/platform-server": "^7.2.7",
-    "@angular/router": "^7.2.7",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ng-bootstrap/ng-bootstrap": "^4.0.4",
     "@ngrx/effects": "^7.3.0",
     "@ngrx/entity": "^7.3.0",
@@ -45,7 +45,7 @@
     "jquery": "^3.4.1",
     "lodash": "^4.17.11",
     "material-design-icons": "^3.0.1",
-    "moment": "^2.23.0",
+    "moment": "^2.21.0",
     "ng-block-ui": "^2.1.5",
     "ng-inline-svg": "^8.3.0",
     "ngrx-store-freeze": "^0.2.4",
@@ -56,19 +56,20 @@
     "ngx-perfect-scrollbar": "^7.2.1",
     "ngx-permissions": "^6.0.5",
     "ngx-toastr": "^10.0.4",
+    "ngx-ui-loader": "^7.2.2",
     "node-sass": "^4.12.0",
     "object-path": "^0.11.4",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.5.3",
     "tslib": "^1.9.0",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.4",
     "@angular/cli": "^7.3.4",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "^7.2.7",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@angular/material": "^7.3.3",
     "@angular/material-moment-adapter": "^7.3.3",
     "@ngrx/schematics": "^7.3.0",
@@ -89,6 +90,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
